Removed ItemIEBase dependance
The ItemIEBase was absolutely unsuitable to inherit from, since it does nothing besides promoting bad coding practice of tying registry names together with translation keys. Minecraft Forge officially states that this kind of relationship is considered a bad coding practice. As such, it was removed, and replaced with much cleaner implementation.